move Streamable, ToJsonDict and FromJsonDict traits into a separate crate. They depends on Error, so move that as well
