+++ chia-protocol/Cargo.toml
@@ -18,6 +18,7 @@ hex = "0.4.3"
 chia_streamable_macro = { version = "0.2.4", path = "../chia_streamable_macro" }
 chia_py_streamable_macro = { version = "0.1.3", path = "../chia_py_streamable_macro", optional = true }
 clvmr = "0.2.7"
+chia-traits = { path = "../chia-traits" }
 
 [dev-dependencies]
 rstest = "0.17.0"
