+++ chia-protocol/src/lib.rs
@@ -1,11 +1,5 @@
-#[cfg(feature = "py-bindings")]
-pub mod from_json_dict;
-#[cfg(feature = "py-bindings")]
-pub mod to_json_dict;
-
 pub mod bls;
 pub mod bytes;
-pub mod chia_error;
 pub mod chia_protocol;
 pub mod classgroup;
 pub mod coin;
@@ -23,7 +17,6 @@ pub mod proof_of_space;
 pub mod reward_chain_block;
 pub mod slots;
 pub mod spend_bundle;
-pub mod streamable;
 pub mod vdf;
 pub mod wallet_protocol;
 pub mod weight_proof;
@@ -47,7 +40,6 @@ pub use crate::proof_of_space::*;
 pub use crate::reward_chain_block::*;
 pub use crate::slots::*;
 pub use crate::spend_bundle::*;
-pub use crate::streamable::*;
 pub use crate::vdf::*;
 pub use crate::wallet_protocol::*;
 pub use crate::weight_proof::*;
