+++ chia-traits/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "chia-traits"
+version = "0.1.0"
+edition = "2021"
+
+[features]
+py-bindings = ["dep:pyo3", "dep:chia_py_streamable_macro"]
+
+[dependencies]
+pyo3 = { version = "0.19.0", features = ["extension-module", "multiple-pymethods"], optional = true }
+chia_py_streamable_macro = { version = "0.1.3", path = "../chia_py_streamable_macro", optional = true }
+chia_streamable_macro = { version = "0.2.4", path = "../chia_streamable_macro" }
+clvmr = "0.2.7"
+sha2 = "0.9.9"
+hex = "=0.4.3"
