+++ wheel/Cargo.toml
@@ -20,5 +20,6 @@ hex = "=0.4.3"
 pyo3 = { version = "=0.19.0", features = ["extension-module", "multiple-pymethods"] }
 chia = { path = "..", features = ["py-bindings"] }
 chia-protocol = { path = "../chia-protocol", features = ["py-bindings"]  }
+chia-traits = { path = "../chia-traits", features = ["py-bindings"]  }
 chia_py_streamable_macro = { path = "../chia_py_streamable_macro" }
 chia_streamable_macro = { path = "../chia_streamable_macro" }
